backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==2.3.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7
bcrypt==4.1.2
python-multipart==0.0.6
boto3==1.34.14
slowapi==0.1.9
openai==1.10.0
python-telegram-bot==20.7
langgraph==0.0.20
langchain==0.1.4
langchain-openai==0.0.5
steel-sdk==0.1.0
imaplib2==3.6
email-validator==2.1.0
sentry-sdk[fastapi]==1.40.0
apscheduler==3.10.4
httpx==0.26.0
pillow==10.2.0
python-dotenv==1.0.0
aiofiles==23.2.1

# ── Test dependencies ──────────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
respx==0.20.2
